feat(car): create list of cars from names
Create a list of cars from names
	- name validation takes place during object creation
	- max len (5) is stored in the Car class
	- non-blank
	- init throws IllegalArgumentException by default

Move user input function into a separate GetUserInput file
	- clean up validation function
	- make validation functions private
